% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_api.R
\name{gpt_api}
\alias{gpt_api}
\title{Wrapper for the Open AI API}
\usage{
gpt_api(
  txt,
  prompt = NULL,
  labels = NULL,
  sep = "|:||:|",
  expertise = NULL,
  certainty = FALSE,
  justification = FALSE,
  model = "gpt-3.5-turbo",
  ...
)
}
\arguments{
\item{txt}{A vector with strings representing text.}

\item{prompt}{A string representing the prompt that is sent to the GPT model. Can be NULL.}

\item{labels}{A vector with n values representing the classes the GPT should annotate the texts with.}

\item{sep}{As GPT only text a string, the labels provided in the vector are collapsed into one string separated by this character. Defaults to " :||: ", which seems to work well and is unlikely to be part of a text.}

\item{expertise}{A string that will be added to the prompt up front that could help to "prompt engineer" the model to do the task better, e.g., "You are an expert identifying topics from newspaper articles."}

\item{certainty}{If TRUE, the output contains a second column with a numerical value expressing GPT's certainty in classifying the text.}

\item{justification}{If TRUE, the output contains another column that includes a short justification for the classification.}

\item{model}{Which model should be used. Defaults to "gpt-3.5-turbo". For a list of available models, see: https://platform.openai.com/docs/models}

\item{output}{A string in which you explain the type of output format you want to get. Default is "basic", which returns in standard data frame (tidy tibble format).}

\item{n_justification}{Maxium length of the justification in words. Defaults to 20.}
}
\description{
Allows to access OPEN AI's GPT-models and run typical text analysis prompts using zero-shot classification.
}
\examples{
\dontrun{
Sys.setenv(OPENAI_API_KEY = 'XXX') # Provide access token, otherwise, it won't work

text = c("To be, or not to be: that is the question.", 
         "An atom is a particle that consists of a nucleus of protons and neutrons surrounded by an electromagnetically-bound cloud of electrons.",
         "Senate passes stopgap bill to avert government shutdown.")

gpt_api(txt = text, 
        labels = c("Poetry", "Politics", "Physics"))


text2 = c("This movie sucks. I really don't like it. Boring as hell, and no story whatsoever!", 
          "This film really struck a nerve for me. Fantastic acting too",
          "This movie is about a physicist who invented the atomic bomb.")

gpt_sentiment(txt = text2)
}
}
