% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf_api.R
\name{hf_api}
\alias{hf_api}
\title{Wrapper for the Huggingface Inference API}
\usage{
hf_api(inputs, url, ...)
}
\arguments{
\item{inputs}{A vector with strings representing text.}

\item{url}{link to a model uploaded to huggingface (e.g., https://api-inference.huggingface.co/models/facebook/bart-large-mnli)}

\item{...}{further parameters that to extract}
}
\description{
Allows to access the Huggingface API and run typical text analysis prompts with huggingface models.
}
\examples{
\dontrun{
Sys.setenv(HF_API_TOKEN = "XXX") # Provide access token, if not only very limited rates are available

text = c("To be, or not to be: that is the question.", 
         "An atom is a particle that consists of a nucleus of protons and neutrons surrounded by an electromagnetically-bound cloud of electrons.",
         "Senate passes stopgap bill to avert government shutdown.")

hf_api(inputs = text, 
       url = "https://api-inference.huggingface.co/models/MoritzLaurer/deberta-v3-large-zeroshot-v1", 
       parameters = list(candidate_labels = c("Poetry", "Politics", "Science")))
}
}
