% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_kwic.R
\name{tidy_kwic}
\alias{tidy_kwic}
\title{Keyword-in-context searches}
\usage{
tidy_kwic(x, keyword, window = 6)
}
\arguments{
\item{x}{A data frame or tibble that contains a column called "text" that contains the text that should be searched for keywords.}

\item{keyword}{a string that represents the keyword}

\item{window}{the number of context words to be displayed around the keyword}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
For a text or a collection of texts stored in a tibble, return a
list of a keyword in its immediate context.
}
