% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_api.R
\name{gpt_api}
\alias{gpt_api}
\title{Wrapper for the Open AI API}
\usage{
gpt_api(
  txt,
  prompt = NULL,
  labels = NULL,
  sep = "|",
  expertise = NULL,
  output = "basic",
  certainty = FALSE,
  justification = FALSE,
  n_justification = 20,
  model = "gpt-3.5-turbo"
)
}
\arguments{
\item{txt}{A vector with strings representing text.}

\item{prompt}{A string representing the prompt that is sent to the GPT model. Can be NULL.}

\item{labels}{A vector with n values representing the classes the GPT should annotate the texts with.}

\item{sep}{As GPT only text a string, the labels provided in the vector are collapsed into one string separated by this character. Defaults to "|", which seems to work well.}

\item{expertise}{A string that will be added to the prompt up front that could help to "prompt engineer" the model to do the task better, e.g., "You are an expert identifying topics from newspaper articles."}

\item{output}{A string in which you explain the type of output format you want to get. Default is "basic", which returns in standard data frame (tidy tibble format).}

\item{certainty}{If TRUE, the output contains a second column with a numerical value expressing GPT's certainty in classifying the text.}

\item{justification}{If TRUE, the output contains another column that includes a short justification for the classification.}

\item{n_justification}{Maxium length of the justification in words. Defaults to 20.}

\item{model}{Which model should be used. Defaults to "gpt-3.5-turbo". For a list of available models, see: https://platform.openai.com/docs/models}

\item{...}{further parameters that to extract}
}
\description{
Allows to access OPEN AI's GPT-models and run typical text analysis prompts using zero-shot classification.
}
\examples{
\dontrun{
Sys.setenv(OPENAI_API_KEY = 'XXX') # Provide access token, otherwise, it won't work

text = c("To be, or not to be: that is the question.", 
         "Come forth into the light of things, Let Nature be your teacher.")

gpt_api(txt = text, 
        labels = c("Poetry", "Politics", "Science"))
}
}
